requirements.txt
Django==4.2.1
djangorestframework
django-cors-headers
names